consolidate duplicated code into a test only crate
