+++ bin/yerdd/Cargo.toml
@@ -76,9 +76,9 @@ nix            = { workspace = true, features = ["user"] }
 
 [dev-dependencies]
 tempfile       = { workspace = true }
-serde_json     = { workspace = true }
 minisign       = { workspace = true }
 tokio          = { workspace = true, features = ["macros", "rt-multi-thread", "net", "io-util", "sync", "time"] }
+yerd-depcheck  = { path = "../../crates/yerd-depcheck" }
 
 [lints]
 workspace = true
