+++ bin/yerdd/src/wordpress_versions.rs
@@ -167,89 +167,7 @@ mod tests {
 
     // ── available_versions: cache / stale-fallback behaviour ────────────────
 
-    use std::sync::Arc;
-
-    use tokio::sync::{Mutex, RwLock as TokioRwLock};
-    use yerd_core::{RouterConfig, SiteRouter, Tld};
-    use yerd_platform::PlatformDirs;
-
-    fn dirs_in(tmp: &std::path::Path) -> PlatformDirs {
-        PlatformDirs {
-            config: tmp.join("c"),
-            data: tmp.join("d"),
-            state: tmp.join("s"),
-            cache: tmp.join("ca"),
-            runtime: tmp.join("r"),
-        }
-    }
-
-    /// Copied verbatim from other modules' test suites (`DaemonState` has no
-    /// `Default`, so every module needing a full instance keeps its own copy).
-    fn state_in(tmp: &std::path::Path) -> DaemonState {
-        let dirs = dirs_in(tmp);
-        let router = SiteRouter::new(RouterConfig::with_tld(Tld::new("test").unwrap()));
-        let ca_path = dirs.data.join("ca.cert.pem");
-        let php_manager = Arc::new(Mutex::new(yerd_php::PhpManager::new(
-            yerd_php::TokioProcessSpawner,
-            yerd_php::SystemClock,
-            yerd_php::io::FastCgiProbe,
-            dirs.clone(),
-            yerd_platform::ActivePortBinder::new(),
-            std::process::id(),
-            std::collections::BTreeMap::new(),
-        )));
-        DaemonState {
-            config: Mutex::new(yerd_config::Config::default()),
-            router: Arc::new(TokioRwLock::new(router)),
-            config_path: dirs.config.join("yerd.toml"),
-            dirs,
-            dns_addr: "127.0.0.1:1053".parse().unwrap(),
-            ca_path,
-            ca_fingerprint: yerd_platform::CaFingerprint::new([0u8; 32]),
-            php_ca_bundle: None,
-            php_updates: TokioRwLock::new(std::collections::HashMap::new()),
-            yerd_update: TokioRwLock::new(Vec::new()),
-            update_snapshot: TokioRwLock::new(None),
-            php_manager,
-            service_manager: Arc::new(Mutex::new(crate::services::new_manager(dirs_in(tmp)))),
-            mail_store: Arc::new(yerd_mail::Store::open(tmp.join("mail")).unwrap()),
-            mail: crate::state::MailRuntime { listening: false },
-            http: yerd_ipc::PortStatus {
-                requested: 80,
-                bound: 8080,
-                fell_back: true,
-            },
-            https: yerd_ipc::PortStatus {
-                requested: 443,
-                bound: 8443,
-                fell_back: true,
-            },
-            redirect_https_port: std::sync::Arc::new(std::sync::atomic::AtomicU16::new(8443)),
-            web_unbound: None,
-            dns_unbound: None,
-            boot_id: 1,
-            started_at: std::time::Instant::now(),
-            shutdown_tx: tokio::sync::watch::channel(false).0,
-            restart_requested: std::sync::atomic::AtomicBool::new(false),
-            detect_cache: Arc::new(crate::detect_cache::DetectCache::new()),
-            watch_dirty: tokio::sync::Notify::new(),
-            dumps: Arc::new(crate::dump_server::DumpStore::new()),
-            shim_reconcile: Mutex::new(()),
-            tunnel_manager: Arc::new(Mutex::new(crate::tunnel::new_manager())),
-            cloudflared_resolution: TokioRwLock::new(None),
-            tool_mutate: Mutex::new(()),
-            tunnel_mutate: Mutex::new(()),
-            php_mutate: Mutex::new(()),
-            jobs: crate::jobs::JobRegistry::default(),
-            reserved_names: Mutex::new(std::collections::HashSet::new()),
-            wordpress_versions: TokioRwLock::new(None),
-            wordpress_login_tokens: Arc::new(crate::wordpress_login::LoginTokenRegistry::new()),
-            wordpress_login_prepend_script: None,
-            wordpress_sites: std::sync::Arc::new(tokio::sync::RwLock::new(
-                std::collections::HashMap::new(),
-            )),
-        }
-    }
+    use crate::test_support::state_in;
 
     struct FakeDl(Result<&'static str, &'static str>);
