+++ crates/yerd-php/Cargo.toml
@@ -24,7 +24,7 @@ nix            = { workspace = true }
 [dev-dependencies]
 tokio          = { workspace = true, features = ["macros", "rt-multi-thread", "net", "sync", "time"] }
 tempfile       = { workspace = true }
-serde_json     = { workspace = true }
+yerd-depcheck  = { path = "../yerd-depcheck" }
 
 [lints]
 workspace = true
