+++ crates/yerd-proxy/Cargo.toml
@@ -24,12 +24,12 @@ rustls         = { workspace = true, features = ["ring"] }
 
 [dev-dependencies]
 yerd-tls       = { path = "../yerd-tls" }
+yerd-depcheck  = { path = "../yerd-depcheck" }
 tokio          = { workspace = true, features = ["macros", "rt-multi-thread", "net", "sync", "time", "io-util"] }
 hyper          = { workspace = true }
 http           = { workspace = true }
 http-body-util = { workspace = true }
 bytes          = { workspace = true }
-serde_json     = { workspace = true }
 tempfile       = { workspace = true }
 rustls         = { workspace = true, features = ["ring"] }
 tokio-rustls   = { workspace = true, features = ["ring"] }
