+++ docs/developer/binaries/yerdd.md
@@ -381,8 +381,9 @@ On non-Unix targets these are no-ops (Windows ACL hardening is a later item); th
 ## Tests & invariants
 
 - **`tests/lifecycle.rs`** (Unix) brings the daemon up against a tempdir `PlatformDirs` and exercises the real socket end-to-end: a `Ping`/`Pong` boot round-trip with clean shutdown; a `Park` round-trip that asserts the discovered child site appears in `ListSites` and the canonical path is persisted to `yerd.toml`; and a `SetSecure` round-trip asserting the parked site stays parked, is persisted under an `[[overrides]]` table, and is **not** promoted to `[[linked]]`.
-- **`tests/no_runtime_deps.rs`** is a dependency-graph invariant: the daemon's default-features runtime graph must not pull `anyhow`, the OpenSSL/native-tls family (`openssl`, `openssl-sys`, `native-tls`, `hyper-tls`, `tokio-native-tls`), or the deprecated `fs2`; and `hyper`/`rustls`/`tokio`/`time` must each resolve to a single version. `webpki-roots` is explicitly allowed (the PHP-install client hits public hosts over rustls). This keeps the install path OpenSSL-free. (`notify` + `notify-debouncer-mini`, added for the watcher, pull none of the forbidden crates.)
+- **`tests/no_runtime_deps.rs`** is a dependency-graph invariant: the daemon's default-features runtime graph must not pull `anyhow`, the OpenSSL/native-tls family (`openssl`, `openssl-sys`, `native-tls`, `hyper-tls`, `tokio-native-tls`), or the deprecated `fs2`; and `hyper`/`rustls`/`tokio`/`time` must each resolve to a single version. `webpki-roots` is explicitly allowed (the PHP-install client hits public hosts over rustls). This keeps the install path OpenSSL-free. (`notify` + `notify-debouncer-mini`, added for the watcher, pull none of the forbidden crates.) The actual graph walk is shared with five other crates/binaries - see [`yerd-depcheck`](../crates/yerd-depcheck).
 - Module unit tests pin the site-scan/override/web-root-detection semantics (`startup`), the `SetWebRoot` dispatch paths (explicit/auto/escape/not-found, in `ipc_server`), the freshness-stamped detection cache (`detect_cache`), exit codes (`error`), cert issue/cache identity (`cert_store`), mode enforcement (`secure_fs`), lock acquisition (`single_instance`), and idempotent tracing (`tracing_init`).
+- **`test_support.rs`** holds daemon-wide test fixtures shared across every module's own `#[cfg(test)]` block: `dirs_in`/`state_in` build a tempdir-rooted `PlatformDirs`/fully-populated `DaemonState` (previously copy-pasted per module - seven near-identical ~80-line copies before consolidation); `sign_manifest` mints a throwaway minisign keypair for tests exercising the signed-`php.json` fetch path. A test needing a non-default `DaemonState` field (e.g. a configured `wordpress_login_prepend_script`) mutates the returned value directly rather than this function growing per-caller parameters.
 
 ## See also
